Add support for xApp registration
This change adds support for xApp registration and deregistration.
Once the xApp starts running it sends messages to the AppMgr to
registrate itself in the RIC.
When the xApp is shutting down it also sends deregistration requests
to the AppMgr.

This change also add signal handlers to the xApp that allow to handle
the SIGINT and SIGTERM signals in order to gracefully shutdown and
release both the local (xApp) and RIC resources.