Add support to subscribe to a given Global gNodeB ID
This change adds support to define a given E2Node ID the xApp
should subscribe. The xApp receives the E2Node ID via command
line, builds the corresponding E2Node ID and subscribes to it
in the RIC platform.
Now a given xApp instance can subscribe/unsubscribe all, or
a given E2Node instance.

This change also adds support for PLMN ID, but for now it is
the constant "747".

This also adds support to load the configuration from the
xApp descriptor file.

Fixed all compilation warnings.