Fix xApp registration and subscription behavior
Current implementation keeps the xApp running even when either the
registration or subscription fails. This does not allow the
Kubernetes know the pod is not working correctly. If either the
registration or subscription fail, then the experiment won't run.

This change shutdown gracefully the xApp when either the registration
or subscription fail. As a result, kubernetes will restart the pod
which will retry to start again.