Add support for REST notification
The xApp did not have support to receive notification via the REST
interface. The Subscription manager is one of the RIC components that
sends such notifications to xApps, and keeps retrying to send those
messages if no HTTP endpoint is available on the corresponding xApp.

This change implements an HTTP listener that handles REST
notifications. Currently, REST notifications are only shown in
the logger console.