Fix subscription notification handler
The subscription notification handler did not process correctly the
message payload when it receives a notification with an unsuccessful
subscription outcome through the REST interface.

This change fixes this issue. The notification handler issues a
SIGTERM signal to shutdown the xApp allowing the Kubernetes to
restart it to retry the subscription.