Add support to subscribe to any E2 NodeB
This xApp only supports to subscribe to gNodeB preventing it to
work with srsRAN which implements an eNodeB. E2 Nodes are fetched
from SDL.

This change implements a generic feature that allows this xApp to
subscribe to any E2 NodeB by using the MCC and MNC to generate the
PLMN ID, which is used along with E2 NodeB ID to select a given
E2 Node to send subscriptions.
Basically, this change allows the xApp to subscribe to all connected
E2 Nodes, or it can subscribe to a single E2 Node based on MCC, MNC,
and E2 Node ID received from command line.

SDL is no longer used to fetch the E2 Nodes since the SDL is not
updated when an E2 Node is down. This change fetches the current
connected E2 Nodes from the E2 Manager.