Add support for E2SM-RC and E2SM-KPM subscriptions
Although both SMs are supported, only E2SM-KPM subscription is enabled
in this change.
NOTE: RIC Indications are not tested, and most likely won't work with
KPM as the xApp sends control messages as receiving an RIC indication, i.e.,
the behavior as receiving an indication message was not changed (RC behavior).